WIP. Исправляет ререндеры главной страницы
