+++ src/pages/main/main.tsx
@@ -1,29 +1,15 @@
 import cn from 'classnames';
-import { useEffect } from 'react';
+import { ReactNode, useState } from 'react';
 
 import HeaderNav from '../../components/header-nav/header-nav';
 import Header from '../../components/header/header';
 import Logo from '../../components/logo/logo';
 
-import { useAppDispatch, useAppSelector } from '../../hooks';
 import TabsList from '../../components/tabs-list/tabs-list';
 import MainContent from '../../components/main-content/main-content';
-import MainEmpty from '../../components/main-empty/main-empty';
-import { selectOffersByCity } from '../../store/offers/offers-selector';
-import { fetchOffers } from '../../store/offers/offers-action';
-import { NameSpace } from '../../const';
 
-function Main(): JSX.Element {
-  const dispatch = useAppDispatch();
-
-  const city = useAppSelector((state) => state[NameSpace.Offers].city);
-  const cityOffers = useAppSelector(selectOffersByCity);
-
-  const offersIsEmpty: boolean = cityOffers.length === 0;
-
-  useEffect(() => {
-    dispatch(fetchOffers());
-  }, [dispatch]);
+function Main(): ReactNode {
+  const [isOffersEmpty, setIsOffersEmpty] = useState<boolean>(true);
 
   return (
     <div className="page page--gray page--main">
@@ -36,22 +22,18 @@ function Main(): JSX.Element {
       <main
         className={
           cn('page__main page__main--index', {
-            'page__main--index-empty': offersIsEmpty
+            'page__main--index-empty': isOffersEmpty
           })
         }
       >
         <h1 className="visually-hidden">Cities</h1>
         <div className="tabs">
           <section className="locations container">
-            <TabsList cityName={city} />
+            <TabsList />
           </section>
         </div>
         <div className="cities">
-          {
-            offersIsEmpty
-              ? <MainEmpty city={city} />
-              : <MainContent city={city} offers={cityOffers} />
-          }
+          <MainContent setIsOffersEmpty={setIsOffersEmpty} />
         </div>
       </main>
     </div>
