+++ src/store/offers/offers-selector.ts
@@ -2,9 +2,13 @@ import { createSelector } from '@reduxjs/toolkit';
 import { State } from '../../types/state';
 import { NameSpace } from '../../const';
 
+const selectCity = (state: State) => state[NameSpace.Offers].city;
+
+const selectDataLoading = (state: State) => state[NameSpace.Offers].loading;
+
 const selectOffersByCity = createSelector(
   [(state: State) => state[NameSpace.Offers].city, (state: State) => state[NameSpace.Offers].data],
   (city, offers) => offers.filter((offer) => offer.city.name === city)
 );
 
-export {selectOffersByCity};
+export {selectOffersByCity, selectCity, selectDataLoading};
